requirements.txt
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
Django==4.2.10
django-cors-headers==4.3.1
django-extensions==3.2.3
django-filter==23.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
factory-boy==3.3.0
Faker==22.5.1
inflection==0.5.1
Markdown==3.5.2
minio==7.2.0
packaging==25.0
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.22
pycryptodome==3.22.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
setuptools==80.4.0
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
